Imporve regex for allowing closing parens to trail DOIs
Regex was improved with a bunch of help from @mudge.

Whereas the first version lead to rogue periods and quotation marks
being added to the end of DOIs, this now only allows ")" to be at the
end of DOIs.

Of course, this still comes with the downside of trapping ) when it
shouldn't.