Separately multiply and sum ISBN digits and weights
This makes the separate steps of the checksum algorithm clearer rather
than squashing them together into a single reduce step. It also makes it
more practical to destructure the zipped pairs and give them meaningful
names, which is an extra aid to clarity.

If/when we start requiring Ruby 2.4, this can use .sum instead of
.reduce(:+).