Support ISBNs separated by Unicode spaces & dashes
Use Unicode character properties to extract ISBNs that have their digits
separated by any Unicode space or dash. While we typically rely on the
caller to clean up their data, the ISBN format permits such use of
whitespace and dash punctuation so we should be permissive in what we
accept.