Only extract ISBNs with consistent hyphenation
As we've seen some false positives in our projects where numbers
separated by a mix of spaces and hyphens are incorrectly extracted as
ISBNs, be a little stricter in ISBN extraction.

Specifically, only permit spaces or dashes between numbers as long as
those are consistent: e.g. an ISBN separated by only spaces or only
dashes but not a mix.

This means that something like 978-0-80-506909-9 is permitted but
978-0 80-506909 9 is not.