Add strict mode to DOI.extract
A very small percentage of DOIs have trailing `.`.
Add an `options` parameter to `DOI.extract`, so that we can change the
default behaviour and extract DOIs while keeping trailing `.`.
After scanning a Crossref mirror, we only found around 5K DOIs ending in
between 1 to 3 `.`. So this was the only scenario we've added to the
regex, to try and keep it as strict as possible.

There's also been a change to the Travis config. This was due to the
release of Bundler 2.0, which dropped support for Ruby versions 2.2 and
older. More can be found here:
https://docs.travis-ci.com/user/languages/ruby/#bundler-20